Use NonZeroUsize inside NodeId
This allows `Option<NodeId>` to be no larger than `usize`,
and saves 32 bytes of memory per node.

To allow IDs to be used directly as indexes,
the `Vec<Node<T>>` storage has an **uninitialized** item at index 0.
Safety-wise, this "infects" the entire crate:
every access to this `Vec` needs to make sure to skip that first (non-) item.